move AbstractUnassembledWebAppMojo and AbstractWebAppMojo from jetty-maven-core to jetty-maven-plugin
