+++ jetty-maven/jetty-maven-plugin/src/main/java/org/eclipse/jetty/maven/plugin/JettyStartMojo.java
@@ -15,7 +15,6 @@ package org.eclipse.jetty.maven.plugin;
 
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.ResolutionScope;
-import org.eclipse.jetty.maven.AbstractUnassembledWebAppMojo;
 
 /**
  *  <p>
